move skills list to skillsData
